Reference

DataCurator.ParallelCounter — Type

Usage
QT = ParallelCount(zeros(Int64, Base.Threads.nthreads()), Int64(0))
QT.data[threadid()] = ...

DataCurator._printtoslack — Method

Print a message to a connected slack

DataCurator.apply_to — Method

apply_to(x, f; base=true)

    Where x is a path, if base=false, return f(x), otherwise works on the last part of the path

DataCurator.apply_to_image — Method

apply_to_image(img, operators::AbstractVector{T}) where {T<:AbstractString}

Apply each of the operators, left to right, to the image (in place.)

Operators can be any unary operators in scope that can be vectorized, e.g. log, sin, cos, abs, ...

DataCurator.bottomup — Method

topdown(node, expander, visitor; context=nothing, inner=expand_sequential)
    Recursively apply visitor onto node, until expander(node) -> []
    If context is nothing, the visitor function gets the current node as sole arguments.
    Otherwise, context is expected to contain: "node" => node, "level" => recursion level.
    Inner is the delegate function that will execute the expand phase. Options are : expand_sequential, expand_threaded

    Traversal is done in a post-order way, e.g. visit after expanding. In other words, leaves before nodes, working from bottom to top.

DataCurator.buildcomp — Method

buildcomp(dataframe, column, operator, value)

Dispatches to the correct form of operator.(df[:,col], value)

operator (String) can be one of:
less, leq, smaller than, more, greater than, equals, euqal, is, geq, isnan, isnothing, ismissing, iszero, <, >, <=, >=, ==

The operator can be negated: ["not", "less"]
Multi-argument comparison are also supported:
- between [x, y]
- in [x, y, z]

You can repeat columns, but non-existing columns are an error.

```julia
df = DataFrame(zeros(2,2),:auto)
df[1,:] .= 5
cols, ops, vals = ["x1", "x1", "x1"], [["not", "in"],"less", ["not", "isnan"]], [[1,2,3,5],10, "NaN"]
@info reduce(&., buildcomp(df, c, o, v) for (c, o, v) in zip(["x1", "x1", "x1"], )
```

DataCurator.canwrite — Method

canwrite(filename)

tests if `filename` can be opened for writing

DataCurator.collapse_functions — Method

collapse_functions(fs; left_to_right=false)

Generalization of (f, g) --> x->(f(g(x))) for any set of functions
left_to_right : g(f(x)), otherwise f(g(x))

DataCurator.column_names — Method

column_names(csvfile)

Return an array of the column names for `csvfile`

DataCurator.config_log — Function

config_log(minlevel=Logging.Debug)

Configure verbose logging. Changes the global logger.

DataCurator.create_template_from_toml — Method

create_template_from_toml(tomlfile)

Parse a toml encoded recipe, decode all the actions and conditions, and return a configuration (Dict) and executable template.

```julia
c, t = create_template_from_toml(tomlfile)
delegate(c, t)
```

DataCurator.decode_dataframe_function — Method

decode_dataframe_function(x::AbstractVector, glob::AbstractDict)

Decodes an entry of the form [command, [(col, op, vals)+]] into a function object for the template

Both the single tuple version
    command, (col, op, vals)
and longer version
    command, [(col, op, vals), ...]
are valid, and dealt by m dispatch.

Note that an entry can be any of:
    - col, op   (for isnan, isnothing, ...)
    - col, op, val (for <, >, ...)
    - col, op, vals (for in, between, ....)

DataCurator.decode_function — Method

decode_function(f::AbstractDict, glob::AbstractDict; condition=false)

Dispatched method for transform entries

DataCurator.delegate — Method

delegate(config, template)
Uses the configuration, and template create by `create_template_from_toml', to execute the verifier as specified.
Returns the counters and file lists, if any are defined.

DataCurator.describe_image — Method

describe_image(x::AbstractArray, axis::Int64)::DataFrame

Describe the array x sliced along axis.

DataCurator.describe_image — Method

describe_image(x, dimension)

Create a DataFrame that contains a succinct description of the intensity distribution of the image.
For example, given a 3D image, with dimension=3, describes the intensity distribution of each z-slice.

DataCurator.describe_image — Method

describe_image(x)

Create a DataFrame that contains a succinct description of the intensity distribution of the image.

DataCurator.describe_objects — Method

describe_objects(img)

For an image, use connected compoments to split the image into objects, and record statistics of each object, return a DataFrame.

DataCurator.describe_objects — Method

describe_objects(img)

For an image, use connected compoments to split the image into objects, and record statistics of each object, return a DataFrame.

DataCurator.dimg — Method

dimg(x)

Describes the image argument as a collapsed Float64 array and returns the statistical moments.

DataCurator.dostep — Method

doset(node, tuple, on_success)

Visitor function, evaluates if(tuple[1](node) -> tuple[2]

Dispatched by type for the condition-action or condition-action-counteraction

DataCurator.dostep — Method

doset(node, tuple, on_success)

Visitor function, evaluates if(tuple[1](node) -> tuple[2] else -> tuple[3]

DataCurator.file_attribute — Method

file_attribute(fname, attribute, value, op)

Checks if the `attribute` (size, modified) of given `fname` matches the given conditions.

# Examples
```julia-repl
julia> file_attribute("test.txt", "size", "10GB", "<")
true
```

DataCurator.filename — Method

filenam(fname)

Return the filename of a file, e.g. "/a/b/c.txt" returns "c.txt".

DataCurator.filepath — Method

filepath(x::AbstractString)

Given a file, return only the path

DataCurator.flipfunctor — Method

flipfunctor(f)

Given a function argument, return its negation.

# Example
```julia-repl
julia> f = x -> false
julia> g=flipfunctor(f)
julia> g(2)
true
```

DataCurator.gaussian — Method

gaussian(img, sigma)

Gaussian blur with σ

DataCurator.generate_counter — Function

Make a count and counting functor that can be incremented by threads
```
c, ct = generate_counter()
ct(something)
@info c # "Counter = 1"
# Threaded version
pc, pct = generate_counter(true; x->reduce(*, size(x))))
a = zeros(3,3,3)
pct(a) # Threadsafe writes
# Printing the counter is not threadsafe, only read when all threads have finished.
@info pc # "Counter = 27"

DataCurator.getextent — Method

For a bounding box, get the XY span (diagonal), Z range, and z center

DataCurator.groupbycolumn — Method

groupbycolumn(dataframe, columns, targets, functions, names)

Group a dataframe by `columns`, selecting `targets`, and using `functions` to do the aggregation, e.g. sum, max, ... .
```julia
dfa = groupbycolumn(df, ["cell", "replicate"], ["area"] ["sum"], ["area_of_sum_per_cell"])
```

DataCurator.groupbycolumn — Method

groupbycolumn(dataframe, columns, targets, functions, names)

Group a dataframe by `columns`, selecting `targets`, and using `functions` to do the aggregation, e.g. sum, max, ... .
```julia
dfa = groupbycolumn(df, ["cell", "replicate"], ["area"] ["sum"], ["area_of_sum_per_cell"])
```

DataCurator.handlecounters! — Method

handlecounters!(entries, key, global_dict)

Decodes user specified counters (file size, count, etc)
Stores the result in global_dict

DataCurator.has_columns_named — Method

has_columns_named(file, names)

Check that `file` is a DataFrames D where `names` ⊂ of the column names(D).

DataCurator.image_colocalization — Function

image_colocalization(directory, window, filter, condition, outputdir)

Compute for a pair of images in `directory` all colocalization metrics.
`condition` controls the types of files to be considered, default 'is_img'
Filter allows you to provide pattern matching, say you want only `1.tif` and `2.tif`, then "[1,2]*.tif" would work
For each metric results are saved in outputdir.

DataCurator.image_files — Method

image_files(dir)

Shorthand for type_files(dir, is_img)

DataCurator.invert — Method

invert(image)

For a normed ([0-1]) image, return  1 - image

DataCurator.is_mesh — Method

is_mesh(x)

Tries to load file `x` as a mesh, false if not possible

DataCurator.keep_going — Method

keep_going(_)

A placeholder function to allow explicit continuing of processing.

DataCurator.laplacian — Method

laplacian(image)

Laplacian of image (2nd derivative of intensity)

DataCurator.load_content — Method

load_content(filename)

Tries to access common formats of content, currently supports:
	tif/png/csv/txt/json/dlp/gsd/mesh formats/mat/hdf5

DataCurator.load_dlp — Method

load_dlp(file)

Load DLP superresolution data

DataCurator.load_gsd — Method

load_gsd(file)

Load GSD superresolution data

DataCurator.load_mesh — Method

load_mesh(x)

Load the contents of a mesh file

DataCurator.load_rainstorm — Method

load_rainstorm(file)

Load Rainstorm data

DataCurator.log_to_file_with_message — Function

log_to_file(x, logfile, reason)

Record file or path `x` to a logfile, with a given message `reason`

DataCurator.mk_remote_path — Function

mk_remote_path(path)

Use SSH to interactively create `path`.
Assumes you have a valid SSH configuration, see `validate_scp_config`
If path == "", tries to create the path you set in ssh config (ssh.json)

DataCurator.parse_acsym — Method

Helper function to parse all functions

DataCurator.read_counter — Method

read_counter(counter)

Sum a parallel or sequential counter where counter.data[threadid()]

DataCurator.reduce_image — Method

reduce_image(image, operator, dimensions)

Apply a reduction (for example maximum projection) along the specified dimension, e.g. 3 for Z.

DataCurator.reduce_images — Method

reduce_images(list, fname::AbstractString, op::AbstractString)

Given list of K-D images (tif), stack to K+1, then apply `op` along K+1 th dimension.
Save to fname in K-D tif

Example
maxproject = (list, fname) -> reduce_images(list, fname, "maximum")

DataCurator.scptmp — Method

scptmp(tempfile, config, path)

Copy tmp using scp to remote path.
Remove `tmp` on success

DataCurator.send_to — Method

/a/b/c, /a/b/c/d/e, /x/y
    if keeprelative
    -> /x/y/c/d/e
    if ~keeprelative
    -> /x/y/e

DataCurator.show_warn_with_message — Method

show_warn_with_message(x, message)

For a matched file or directory `x`, log a warning with `x` and a custom message.

DataCurator.slice_image — Method

slice_image(img::Array, dim, m::T, M::T)

Slice the image along dimension `dim` from m to M

DataCurator.subdirs — Method

subdirs(path)

Return an array of sub-directories for path

DataCurator.threshold_image — Method

threshold_image(img, operator, value)

Returns the image where all pixels are set to zero where operator(pixel, value) == true

```julia
threshold_image(img, "<", 0.25)
```

DataCurator.threshold_image — Method

treshold(image, operator, value)

Set the image to zero where operator(image, value) == true.
Operator can be one of '<', '>', 'abs >', 'abs <'.

DataCurator.tmpcopy — Method

tmpcopy(x; seed=0, length=40)

Create a temporary copy of file x (with same extension), of length 40.
40 means there's a 1/10000 for a collision of 2 identical files.
Note that this function is used on executing a template, in parallel, so it's not the total of number files being processed, but the number of files being processed in the same window of time.
Return the temporary file name

DataCurator.topdown — Method

topdown(node, expander, visitor; context=nothing, inner=expand_sequential)
Recursively apply visitor onto node, until expander(node) -> []
If context is nothing, the visitor function gets the current node as sole arguments.
Otherwise, context is expected to contain: "node" => node, "level" => recursion level.
Inner is the delegate function that will execute the expand phase. Options are : expand_sequential, expand_threaded

Traversal is done in a pre-order way, e.g. visit before expanding.

DataCurator.transform_copy — Method

transform_copy(x, f)
    x' = f(x) for a file or directory, a copy rather than a move. Refuses to act if x' exists.

DataCurator.transform_inplace — Method

transform_inplace(x, f)
    x = f(x) for a file or directory. Refuses to act if x' exists.

DataCurator.try_mesh — Method

try_mesh(x)

Try to load the contents of a mesh file, return nothing if it fails

DataCurator.type_files — Method

type_files(dir, condition)

Find all files in `dir` matching condition.

Example: `type_files(dir, is_2d_img)`

DataCurator.upload_to_owncloud — Method

upload_to_owncloud(name)

DataCurator.upload_to_scp — Method

upload_to_scp(file)

For a given file, upload to a remote server using SCP.
Assumes `validate_scp_config` has succeded.

*Notes*
- a file with whitespace in name or path, will have this converted to replace all whitespace with '_'
- copying works asynchronously, given that the copy operation can run for a long time, and so avoids blocking

DataCurator.validate_scp_config — Method

validate_scp_config(configfile)

Try to load SCP config from JSON.
If successful, returns true, and saves configuration in ENV["DC_SSH_CONFIG"]

DataCurator.verifier — Method

verifier(node, templater::Dict, level::Int)
Dispatched function to verify at recursion level with conditions set in templater[level] for node.
Will apply templater[-1] as default if it's given, else no-op.

DataCurator.verifier — Method

verifier(node, template::Vector, level::Int)
Dispatched function to verify at recursion level with conditions set in template for node.
Level is ignored for now, except to debug

DataCurator.verify_dispatch — Method

verify_dispatch(context)
Use multiple dispatch to call the right function verifier.

DataCurator.verify_template — Method

verify_template(start, template; expander=expand_filesystem, traversalpolicy=bottomup, parallel_policy="sequential")
Recursively verifies a dataset anchored at start using a given template.
For example, start can be the top directory of a filesystem.
A template has one of 2 forms:
    - template = [(condition, action_on_fail), (condition, action), ...]
        - where condition accepts a node and returns true if ok, false if not.
        - action is a function that accepts a node as argument, and is trigger when condition fails, example warn_on_fail logs a warning
Traversalpolicy is bottomup or topdown. For modifying actions bottomup is more stable.
Parallel_policy is one of "sequential" or "parallel". While parallel execution can be a lot faster, be very careful if your actions share global state.

DataCurator.wrap_transform — Method

wrap_transform(x::AbstractString)

For file x, generate a temp copy before aggregation on a file list