Reference

DataCurator.ParallelCounter — Type

Usage
QT = ParallelCount(zeros(Int64, Base.Threads.nthreads()), Int64(0))
QT.data[threadid()] = ...

DataCurator.apply_to — Method

apply_to(x, f; base=true)

    Where x is a path, if base=false, return f(x), otherwise works on the last part of the path

DataCurator.bottomup — Method

topdown(node, expander, visitor; context=nothing, inner=expand_sequential)
    Recursively apply visitor onto node, until expander(node) -> []
    If context is nothing, the visitor function gets the current node as sole arguments.
    Otherwise, context is expected to contain: "node" => node, "level" => recursion level.
    Inner is the delegate function that will execute the expand phase. Options are : expand_sequential, expand_threaded

    Traversal is done in a post-order way, e.g. visit after expanding. In other words, leaves before nodes, working from bottom to top.

DataCurator.buildcomp — Method

buildcomp(dataframe, column, operator, value)

Dispatches to the correct form of operator.(df[:,col], value)

operator (String) can be one of:
less, leq, smaller than, more, greater than, equals, euqal, is, geq, isnan, isnothing, ismissing, iszero, <, >, <=, >=, ==

The operator can be negated: ["not", "less"]
Multi-argument comparison are also supported:
- between [x, y]
- in [x, y, z]

You can repeat columns, but non-existing columns are an error.

```julia
df = DataFrame(zeros(2,2),:auto)
df[1,:] .= 5
cols, ops, vals = ["x1", "x1", "x1"], [["not", "in"],"less", ["not", "isnan"]], [[1,2,3,5],10, "NaN"]
@info reduce(&., buildcomp(df, c, o, v) for (c, o, v) in zip(["x1", "x1", "x1"], )
```

DataCurator.collapse_functions — Method

collapse_functions(fs; left_to_right=false)

Generalization of (f, g) --> x->(f(g(x))) for any set of functions
left_to_right : g(f(x)), otherwise f(g(x))

DataCurator.decode_dataframe_function — Method

decode_dataframe_function(x::AbstractVector, glob::AbstractDict)

Decodes an entry of the form [command, [(col, op, vals)+]] into a function object for the template

Both the single tuple version
    command, (col, op, vals)
and longer version
    command, [(col, op, vals), ...]
are valid, and dealt by m dispatch.

Note that an entry can be any of:
    - col, op   (for isnan, isnothing, ...)
    - col, op, val (for <, >, ...)
    - col, op, vals (for in, between, ....)

DataCurator.decode_function — Method

decode_function(f::AbstractDict, glob::AbstractDict; condition=false)

Dispatched method for transform entries

DataCurator.delegate — Method

delegate(config, template)
Uses the configuration, and template create by `create_template_from_toml', to execute the verifier as specified.
Returns the counters and file lists, if any are defined.

DataCurator.describe_image — Method

describe_image(x::AbstractArray, axis::Int64)::DataFrame

Describe the array x sliced along axis.

DataCurator.generate_counter — Function

Make a count and counting functor that can be incremented by threads
```
c, ct = generate_counter()
ct(something)
@info c # "Counter = 1"
# Threaded version
pc, pct = generate_counter(true; x->reduce(*, size(x))))
a = zeros(3,3,3)
pct(a) # Threadsafe writes
# Printing the counter is not threadsafe, only read when all threads have finished.
@info pc # "Counter = 27"

DataCurator.parse_acsym — Method

Helper function to parse all functions

DataCurator.read_counter — Method

read_counter(counter)

Sum a parallel or sequential counter where counter.data[threadid()]

DataCurator.reduce_images — Method

reduce_images(list, fname::AbstractString, op::AbstractString)

Given list of K-D images (tif), stack to K+1, then apply `op` along K+1 th dimension.
Save to fname in K-D tif

Example
maxproject = (list, fname) -> reduce_images(list, fname, "maximum")

DataCurator.send_to — Method

/a/b/c, /a/b/c/d/e, /x/y
    if keeprelative
    -> /x/y/c/d/e
    if ~keeprelative
    -> /x/y/e

DataCurator.tmpcopy — Method

tmpcopy(x; seed=0, length=40)

Create a temporary copy of file x (with same extension), of length 40.
40 means there's a 1/10000 for a collision of 2 identical files.
Note that this function is used on executing a template, in parallel, so it's not the total of number files being processed, but the number of files being processed in the same window of time.
Return the temporary file name

DataCurator.topdown — Method

topdown(node, expander, visitor; context=nothing, inner=expand_sequential)
Recursively apply visitor onto node, until expander(node) -> []
If context is nothing, the visitor function gets the current node as sole arguments.
Otherwise, context is expected to contain: "node" => node, "level" => recursion level.
Inner is the delegate function that will execute the expand phase. Options are : expand_sequential, expand_threaded

Traversal is done in a pre-order way, e.g. visit before expanding.

DataCurator.transform_copy — Method

transform_copy(x, f)
    x' = f(x) for a file or directory, a copy rather than a move. Refuses to act if x' exists.

DataCurator.transform_inplace — Method

transform_inplace(x, f)
    x = f(x) for a file or directory. Refuses to act if x' exists.

DataCurator.verifier — Method

verifier(node, templater::Dict, level::Int)
Dispatched function to verify at recursion level with conditions set in templater[level] for node.
Will apply templater[-1] as default if it's given, else no-op.

DataCurator.verify_dispatch — Method

verify_dispatch(context)
Use multiple dispatch to call the right function verifier.

DataCurator.verify_template — Method

verify_template(start, template; expander=expand_filesystem, traversalpolicy=bottomup, parallel_policy="sequential")
Recursively verifies a dataset anchored at start using a given template.
For example, start can be the top directory of a filesystem.
A template has one of 2 forms:
    - template = [(condition, action_on_fail), (condition, action), ...]
        - where condition accepts a node and returns true if ok, false if not.
        - action is a function that accepts a node as argument, and is trigger when condition fails, example warn_on_fail logs a warning
Traversalpolicy is bottomup or topdown. For modifying actions bottomup is more stable.
Parallel_policy is one of "sequential" or "parallel". While parallel execution can be a lot faster, be very careful if your actions share global state.

DataCurator.wrap_transform — Method

wrap_transform(x::AbstractString)

For file x, generate a temp copy before aggregation on a file list